REFACTOR: Allow fractional CPU in resource limits
- Changed the `CPU` field in `ResourceLimits` struct from `int` to `float64` to support fractional CPU values.
- Updated the example in README.md to reflect the change to fractional CPU.
- Added new tests to `sqlite_test.go` to ensure that both fractional and integer CPU values can be unmarshaled correctly.

+++ internal/storage/sqlite.go
@@ -80,7 +80,7 @@ func (a *SourceInfo) Scan(value interface{}) error {
 }
 
 type ResourceLimits struct {
-	CPU      int `json:"cpu"`
+	CPU      float64 `json:"cpu"`
 	MemoryMB int `json:"memoryMB"`
 }
